+++ src/App.js
@@ -1,18 +1,23 @@
 import './App.css';
 import {BrowserRouter as Router, Routes, Route} from 'react-router-dom'
+
 import Login from './pages/Login';
 import Manager from './pages/Manager';
 import Cashier from './pages/Cashier';
 
+import {Provider as ExecutiveProvider} from './context/Executives/ExecutivesContext'
+
 function App() {
   return (
-    <Router>
-      <Routes>
-        <Route exact path='/' element={<Login/>}/>
-        <Route exact path='/manager/*' element={<Manager/>}/>
-        <Route exact path='/cashier/*' element={<Cashier/>}/>
-      </Routes>
-    </Router>
+    <ExecutiveProvider>
+      <Router>
+        <Routes>
+          <Route exact path='/' element={<Login/>}/>
+          <Route exact path='/manager/*' element={<Manager/>}/>
+          <Route exact path='/cashier/*' element={<Cashier/>}/>
+        </Routes>
+      </Router>
+    </ExecutiveProvider>
   );
 }
