+++ src/pages/Manager.js
@@ -4,7 +4,7 @@ import {BrowserRouter as Router, Routes, Route} from 'react-router-dom'
 import Dashboard from './Dashboard'
 import Accounts from './Accounts'
 import Transfer from './Transfer'
-import Executives from './Executives'
+import Executives from './Executives/Executives'
 import Areas from './Areas'
 import Comissions from './Comissions'
 import Concepts from './Concepts'
